Relocate function
