+++ packages/doenetml/src/Viewer/DocViewer.tsx
@@ -1412,110 +1412,3 @@ class ErrorBoundary extends React.Component<ErrorProps, ErrorState> {
         return this.props.children;
     }
 }
-
-// TODO: fix this function as it assume conventions that are no longer valid
-export function getURLFromRef({
-    cid,
-    activityId,
-    variantIndex,
-    edit,
-    hash,
-    givenUri,
-    targetName = "",
-    linkSettings,
-    search = "",
-    id = "",
-}: {
-    cid?: string;
-    activityId?: string;
-    variantIndex?: number;
-    edit?: boolean;
-    hash?: string;
-    givenUri?: string;
-    targetName?: string;
-    linkSettings: Record<string, any>;
-    search?: string;
-    id?: string;
-}) {
-    // possible linkSettings
-    // - viewURL
-    // - editURL
-    // - useQueryParameters
-
-    let url = "";
-    let targetForATag: string | null = "_blank";
-    let haveValidTarget = false;
-    let externalUri = false;
-
-    if (cid || activityId) {
-        if (cid) {
-            if (linkSettings.useQueryParameters) {
-                url = `cid=${cid}`;
-            } else {
-                // TODO: make this URL work for create another URL to reference by cid
-                url = `/${cid}`;
-            }
-        } else {
-            if (linkSettings.useQueryParameters) {
-                url = `doenetId=${activityId}`;
-            } else {
-                url = `/${activityId}`;
-            }
-        }
-        if (variantIndex) {
-            // TODO: how to specify variant if don't useQueryParameters
-            if (linkSettings.useQueryParameters) {
-                url += `&variant=${variantIndex}`;
-            }
-        }
-
-        if (linkSettings.useQueryParameters) {
-            let baseUrl =
-                edit == true ? linkSettings.editURL : linkSettings.viewURL;
-            if (baseUrl.includes("?")) {
-                if (baseUrl[baseUrl.length - 1] !== "?") {
-                    baseUrl += "&";
-                }
-            } else {
-                baseUrl += "?";
-            }
-            url = baseUrl + url;
-        } else {
-            if (edit == true) {
-                url = linkSettings.editURL + url;
-            } else {
-                url = linkSettings.viewURL + url;
-            }
-        }
-
-        haveValidTarget = true;
-
-        if (hash) {
-            url += hash;
-        } else {
-            if (targetName) {
-                url += "#" + cesc(targetName);
-            }
-        }
-    } else if (givenUri) {
-        url = givenUri;
-        if (
-            url.substring(0, 8) === "https://" ||
-            url.substring(0, 7) === "http://" ||
-            url.substring(0, 7) === "mailto:"
-        ) {
-            haveValidTarget = true;
-            externalUri = true;
-        }
-    } else {
-        url = search;
-
-        let firstSlash = id.indexOf("\\/");
-        let prefix = id.substring(0, firstSlash);
-        url += "#" + prefix;
-        url += cesc(targetName);
-        targetForATag = null;
-        haveValidTarget = true;
-    }
-    return { targetForATag, url, haveValidTarget, externalUri };
-}
